+++ src/not_use/add_z.c
@@ -6,15 +6,15 @@
 /*   By: jvigny <jvigny@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/25 13:07:43 by jvigny            #+#    #+#             */
-/*   Updated: 2023/01/25 14:39:53 by jvigny           ###   ########.fr       */
+/*   Updated: 2023/02/07 15:27:08 by jvigny           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "fdf.h"
 
-void	add_z(t_coordonnee_3d *tab, size_t len)
+void	add_z(t_coordonnee_3d *tab, unsigned int len)
 {
-	size_t	i;
+	unsigned int	i;
 
 	i = 0;
 	while (i < len)
